Create FXBacktestingPython
Algo trading forex – backtesting and optimizing an FX moving average strategy in Python. The crossing over of short and long term moving averages is a well-known signal used in algo trading. The idea is that when the short term average rises above the long term average, it could indicate the price is beginning to rise. Similarly, when the short term average falls below the long term average, it could indicate the price is beginning to fall. See accompanying article at https://mathematicsconsultants.com/2022/10/23/algo-trading-forex-backtesting-and-optimizing-an-fx-moving-average-strategy/.